xmonadify
This involves adding functionality to Main.hs where if it is given an argument, it compiles the argument and runs it.

Presumably the file is a valid haskell file which contains something like 'autoprocMain $ gwernRules' (all the libs are exported so you can write your own Autoproc.Rules.Gwern module even). If there is no argument, we default to 'autoprocMain $ dagitRules', which is a little silly since who but lispy needs his procmail config, but it demonstrates the idea.